+++ experimental/h/hong_kong/source/help/hong_kong.php
@@ -1,0 +1,227 @@
+<?php 
+  $pagename = 'hong kong Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+  <p>Hong Kong keyboard for the Script of National Language for Cantonese of Hong Kong and Canton city accent</p> 
+  <p>Oạt Yữ Quóc Yữ Chị Hương Cỏng Thùng Quỏng Châu Yâm粵語國語字香港同廣州音</p> 
+  <p>Bộ gõ Chữ Quốc Ngữ tiếng Quảng Đông giọng Quảng Châu và Hương Cảng</p>
+
+   <p>粵語國語字係一種拉丁化嘅粵語文字，意在發展粵語表音文字，縮短粵文學習同書寫時間，增強粵語使用者之間嘅凝聚力同自我認同 </p>
+   <p> Oạt-Yữ Quóc-Ngữ-Chị hệ yât chủng Lai-đinh-phá ké Oạt-Yữ mần-chị, ý-chọi phát-chiển Oạt-Yữ biểu-yâm mần-chị, suc-đoản Oạt-Mần học-chập thùng sư-sẻ sị-kan, chăng-khường Oạt-Yữ sỉ-yụng-chẻ chi-kan ké yình-chụi-lịch thùng chị-ngõ yịnh-thùng</p>
+   <p> Chữ Quốc Ngữ tiếng Quảng Đông là một loại chữ viết La-tinh hóa của tiếng Quảng Đông, nhằm mục đích phát triển chữ viết biểu âm của tiếng Quảng Đông, giảm bớt thời gian học và viết tiếng Quảng Đông, nâng cao sức hướng tâm và ý thức về cá tính mình của những người sử dụng tiếng Quảng Đông</p>
+  <p> The script of national language for Cantonese is a romanized Cantonese script aimed at developing a phonetic writing system for Cantonese, reduce the time of learning and writing of Cantonese, while strengthening the cohesion and self-identity among Cantonese speakers.</p>
+   <p>
+
+  <h1>音韻表 phonologic map</h1>
+  <p> 建議以下嘅粵語國語字音韻表打字<p>
+  <p> The suggested form of spelling of Cantonese Romanization<p></p>
+  
+  <h1>韻首 Phụ Âm consonants</h1>
+  <img src='consonants.png' alt='Keyboard Layout'>
+  <p>位於字母a â ă u ư o ô ơ 之前嘅韻首k可以轉變為c</p>
+  <p>The consonant "k" before vowels a â ă u ư o ô ơ can be transformed to "c"</p>
+  <h1>韻腹 Nguyên Âm và Phụ Âm Cuối vowels and end consonants</h1> 
+  <img src='vowels.png' alt='Keyboard Layout'>
+  <p>"iu" "im" "in" "ip" "it"組合用於取代"iêu" "iêm" "iên" "iêp" "iêt"當用嚟表示發音或者串唔係粵語詞源或者唔係漢越詞詞源嘅詞語，譬如話來自嘅英文嘅借詞「𨋢lip」「天拿水thin-nà sủi」，或者係借用漢字嚟表示詞語嘅發音但係字嘅意思同漢字本身嘅意思無關嘅字，譬如話「點解đỉm-cải」</p>
+  <p>The "iu" "im" "in" "ip" "it" combiniation can be used to replace "iêu" "iêm" "iên" "iêp" "iêt" when representing sounds and spelling words without Cantonese origin and without sino-cantonese origin, for example loanword from English, such as "lip" and "thin-nà sủi", or in case of using chinese characters to represent sounds or a word but the meaning of the character is not linked with its origin meaning, such as "點解đỉm-cải"</p>
+   <p>ê同ay用嚟表示一於嘅發音，ê用嚟拼寫漢字詞源嘅字同有明確嘅詞源喺古音入面讀呢個韻腹嘅詞語，ay用嚟拼寫少部分嘅非漢字嘅詞語或者係漢字但係係由不規則音變產生呢個讀音嘅字，譬如話「嚟lày」「睇thảy」「咪may」「咪mạy」「米may」「呢nay」「揮phay」「僞ngạy」「魏ngày」「危ngày」「矮ảy」「哎áy」「曳yay」「喺hảy」「使sảy」「费pháy」「沸pháy」</p>
+ <p>"ê" and "ay" are used to represent same pronunciation. "ê" is used to spell sino-cantonese words and the words with clear etymology that pronounce this vowel in ancient time. "ay" is used to spell few non-sino-cantonese words and sino-cantonese words with irregular sound change, such as「嚟lày」「睇thảy」「咪may」「咪mạy」「米may」「呢nay」「揮phay」「僞ngạy」「魏ngày」「危ngày」「矮ảy」「哎áy」「曳yay」「喺hảy」「使sảy」「费pháy」「沸pháy」</p>
+ <p>如果一個單獨嘅yi音嘅音節聲調係第1、2、3聲調，咁唔使寫後邊嘅i，譬如話「衣y」「倚ỷ」「意ỷ」</p>
+ <p>If a "yi" sound syllable has first, second, or third tone, the "i" behind can be removed, such as 「衣y」「倚ỷ」「意ỷ」</p>
+ <p>如果一個單獨嘅yi音嘅音節聲調係第4、5、6聲調，咁要寫yi，譬如話「移yì」「以yĩ」「易yị」</p>
+ <p>If a "yi" sound syllable has forth, fifth, or sixth tone, it should be spelled as "yi", such as 「移yì」「以yĩ」「易yị」</p>
+   <p>如果一個單獨嘅yư音嘅音節聲調係第1、2、3聲調，咁唔使寫前邊嘅y，譬如話「於ư」「傴ử」「淤ứ」</p>
+ <p>If a "yư" sound syllable has first, second, or third tone, the "i" behind can be removed, such as 「於ư」「傴ử」「淤ứ」</p>
+ <p>如果一個單獨嘅yư音嘅音節聲調係第4、5、6聲調，咁要寫yư，譬如話「輿yừ」「與yữ」「預yự」</p>
+ <p>If a "yư" sound syllable has forth, fifth, or sixth tone, it should be spelled as "yi", such as 「輿yừ」「與yữ」「預yự」</p>
+ <p>如果一個單獨嘅uy音嘅音節聲調係第4、5、6聲調，咁可以寫成vy，譬如話「為vỳ」「偉vỹ」「位vỵ」</p>
+ <p>If a "vy" sound syllable has forth, fifth, or sixth tone, it can be spelled as "vy", such as 「為vỳ」「偉vỹ」「位vỵ」</p>
+ <p>如果一個單獨嘅yiêt, yiên, yiêm音嘅音節聲調係第1、2、3聲調，咁唔使寫中間嘅i，譬如話「噎yết」「煙yên」「腌yêm」</p>
+ <p>If a "yiêt" "yiên" "yiêm" sound syllable has first, second, or third tone, the "i" in the middle can be removed, such as 「噎yết」「煙yên」「腌yêm」</p>
+ <p>如果"y" "ê" 嘅元音前邊冇韻首單獨出現，咁要寫"ơi" "ay"，譬如話「誒ợi」「哎áy」</p>
+ <p>If "y" and "ê" vowels appear without a consonant before, it should be spelled as "ơi" "ay", such as 「誒ợi」「哎áy」</p>
+ <p>如果"â"嘅元音後邊冇韻尾單獨出現，咁要寫"ơ"，譬如話「㗎cớ」「嘞lớ」</p>
+ <p>If "â" vowels appear without a consonant behind, the vowel should be spelled as "ơ", such as 「㗎cớ」「嘞lớ」</p>
+  <h1>音調 Thanh Điệu tones</h1>
+  <img src='tones.png' alt='Keyboard Layout'>
+
+
+</p>
+<p>Specific Alphabet mẫu tự đặc biệt 特別字母<p>
+<table>
+  <tbody>
+  <tr>
+    <th>Character</th>
+    <th>Key pressed</th>
+    <th>Sample input</th>
+    <th>Sample output</th>
+  </tr>
+  <tr>
+    <td>ă</td>
+    <td>aw</td>
+    <td>chac<strong>w</strong></td>
+    <td>chăc則</td>
+  </tr>
+  <tr>
+    <td>â</td>
+    <td>aa</td>
+    <td>yat<strong>a</strong></td>
+    <td>yât一</td>
+  </tr>
+  <tr>
+    <td>ê</td>
+    <td>ee</td>
+    <td>le<strong>e</strong>c</td>
+    <td>lêc叻</td>
+  </tr>
+    <tr>
+    <td>iê</td>
+    <td>ii</td>
+    <td>th</strong>i<strong>i</strong>n</td>
+    <td>thiên天</td>
+    <tr>
+      <td>ô</td>
+      <td>oo</td>
+      <td>ddo<strong>o</strong></td>
+      <td>đô都</td>
+    </tr>
+  <tr>
+    <td>ơ</td>
+    <td>ow</td>
+    <td>co<strong>w</strong></td>
+    <td>cơ㗎</td>
+  </tr>
+  </tbody>
+</table>
+
+<p>tone rules quy tắc thanh điệu 音調規則<p>
+  <p> 對應粵拼嘅第一聲唔使打，第二聲打r，第三聲打s，第四聲打f，第五聲打x，第六聲打j <p>
+  <p> tone 1 for nothing, tone 2 = r, tone 3 =s, tone 4 = f, tone 5 = x, tone 6 = j <p>
+<h2>Tonal diacritics</h2>
+
+<table>
+  <tbody>
+    <tr>
+      <th>Tone</th>
+      <th>Keys added to syllable</th>
+      <th>Sample input</th>
+      <th>Sample output
+      </th>
+    </tr>
+    <tr>
+      <td>1(high level)</td>
+      <td></td>
+      <td>ddoo<strong></strong></td>
+      <td>đô都</td>
+    </tr>
+    <tr>
+      <td>2 (high rising)</td>
+      <td>r</td>
+      <td>cho<strong>r</strong></td>
+      <td>chỏ左
+       </td>
+    </tr>
+    <tr>
+      <td>3 (mid level)</td>
+      <td>s</td>
+      <td>huw<strong>s</strong></td>
+      <td>húi去
+      </td>
+    </tr>
+    <tr>
+      <td>4 (low falling)</td>
+      <td>f</td>
+      <td>tra<strong>f</strong></td>
+      <td>xà茶
+      </td></tr>
+    <tr>
+      <td>5 (low rising)</td>
+      <td>x</td>
+      <td>ma<strong>x</strong></td>
+      <td>mã馬
+      </td>
+    </tr>
+    <tr>
+      <td>6 (low falling)</td>
+      <td>j</td>
+      <td>hee<strong>j</strong></td>
+      <td>hệ係
+      </td>
+    </tr>
+    <tr>
+      <td>7 (high entering)</td>
+      <td>/</td>
+      <td>yaat</td>
+      <td>yât一
+      </td>
+    </tr>
+    <tr>
+      <td>8 (middle entering)</td>
+      <td>s</td>
+      <td>phat<strong>s</strong></td>
+      <td>phát發
+      </td>
+    </tr>
+    <tr>
+      <td>9 (low entering)</td>
+      <td>j</td>
+      <td>phaat<strong>j</strong></td>
+      <td>phật罰
+      </td>
+    </tr>
+  </tbody>
+</table>
+
+<p>examples: 舉例： <p>
+  <p> ar > ả, as > á, af > à, ax > ã, aj > ạ <p>
+  <p> 譬如話：‘兩’ 嘅羅馬字係‘lưỡng’， 咁輸入內容就係‘luongwx’<p>
+
+<h2>Punctuation and Capitalization標點同大小寫biêu điểm thùng đại siểu sẻ</h2>
+<p>A space " " should be added betweem each words or between each syllable of cantonese origin or sino-cantonese origin words.</p>
+<p>每個詞語或者每個粵語詞語或者漢粵詞嘅每個音節之間應該添加空格" "</p>
+<p>The first letter in one sentence should be capitalized</p>
+<p>每個句子嘅第一個字母要大寫</p>
+<p>The first letter in a proper noun such as the name of a particular person and place should be capitalized</p>
+<p>專有名詞譬如話人同地方嘅第一個字母要大寫</p>
+<p>The first letter in each syllable of a Cantonese origin or Sino-cantonese origin proper noun such as the name of a particular person and place should be capitalized</p>
+<p>粵語或者漢粵詞詞源嘅專有名詞嘅每一個音節嘅第一個字母要大寫</p>
+<p>In an acronym, the first letter of each syllable should be capitalized, and a period "." can be added behind each capitalized letter</p>
+<p>係縮寫中，每個音節嘅第一個字母要大寫，每個大寫字母嘅後邊可以添埋一個句號"."</p>
+<p>In a multi-syllable single word, the space " " should be added between each syllable.</p>
+<p>係一個多音字單詞中，每個音節之間應該加入一個連字符"-"</p>
+<p>In a multi-syllable combined word or a Cantonese origin or Sino-cantonese origin word, the hyphen "-" can be added between each syllable for avoid confusion</p>
+<p>係一個多音節嘅組合詞或者粵語詞或者漢粵詞嘅每一個音節之間，可以加入一個連字符"-"避免撈亂</p>
+<p>Other punctuation works similar with punctuations in English and Vietnamese</p>
+<p>其他標點嘅作用同英文同越南文類似</p>
+<p></p>
+  <p> doubletaping leads to generate original letters:  <p>
+  <p> 雙擊音調鍵同特殊轉化鍵，取消轉化，變返原本嘅拼寫 <p>  
+    <p>
+      If more than one tone marking key is pressed, the last one will be used. For example, typing <code>asf</code> will return "à". To write a tone marking key as a normal character, one has to press it twice: <code>her</code> becomes <code>hẻ</code>, while <code>herr</code> becomes <code>her</code>.
+    </p>
+  <p> tap Z key to remove tonations  <p>
+  <p> 使用Z鍵取消音調轉換 <p>
+
+    <a href="https://github.com/Cathaylaboratory/quokyuzi/blob/main/quocyuchiquongchauyam.dic.yaml">Dictionary for Cantonese Quóc Yữ Chị粵語國語字字典香港同廣州音</a>
+    <p></p>
+    <a href="https://github.com/Cathaylaboratory/quokyuzi/blob/main/oatnguquocnguti.yaml">Dictionary for the script of National Language universal version Quóc Ngữ Tị大粵國音字匯－粵語國語字字典總合通用版</a>
+<p></p>
+    <a href="https://t.me/+ax2H-iql9rplYzBk">Cantonese National Language Script Telegram Group粵語國語字谷Oạt Ngữ Quóc Ngữ Tị Cuc</a>
+<!-- <p>See <a href="file:taigi_poj.pdf">this document</a> for all the key combinations.</p> -->
+
+
+<p>範文：南國山河</p>
+<p>Nàm Quóc San Hò</p>
+<p>南國山河南帝居</p>
+<p>Nàm Quóc San Hò Nàm Đế Cui</p>
+<p>截然定分在天書</p>
+<p>Chiệt Yiền Định Phân chọi Thiên Sư</p>
+<p>如何逆虜來侵犯</p>
+<p>Yừ Hò Yịch Lỗ Lòi Xâm Phạn</p>
+<p>汝等行看取敗虛</p>
+<p>Yữ Đẳng Hành Hón Xủi Bại Hui</p>
+
+<p>© cathaylab</p>