+++ release/sil/sil_busa/HISTORY.md
@@ -1,5 +1,9 @@
-Busa Change History
+Busa (SIL) Change History
 =======================
+
+1.5.2 (2025-06-09)
+------------------
+* Update font to new version of Charis
 
 1.5.1 (2025-02-20)
 ------------------