+++ release/sil/sil_mali_azerty/HISTORY.md
@@ -1,6 +1,10 @@
 Clavier du Mali (Azerty) Keyboard Change History
 =======================
 
+3.4.2 (10 June 2025)
+------------------
+* Update font to Charis Mali
+
 3.4.1 (20 Feb 2025)
 ------------------
 * Put OSKFont from .kps file into .kvks file