+++ release/v/vietnam/HISTORY.md
@@ -1,6 +1,34 @@
 vietnam Change History
 ====================
 
+2.0 (2025-07-03)
+----------------
+* Updated by cathaylab
+* significantly shortened the codes
+* deleted "iun" + "n" > "ương" rule
+* fixing several bugs in generation
+* improved the generation order of double pressing "n" for generate "nh", to avoid unnecessary generation.
+
+1.0.8 (2025-05-30)
+----------------
+* Updated by cathaylab
+* fixing several bugs of keyboard codes
+
+1.0.7 (2025-04-20)
+----------------
+* Updated by cathaylab
+* fixing several bugs of keyboard codes
+
+1.0.6 (2025-04-11)
+----------------
+* Updated by cathaylab
+* fixing several bugs in generation
+
+1.0.5 (2025-03-31)
+----------------
+* Updated by cathaylab
+* fixing several bugs in generation
+
 
 1.0.4 (2025-03-20)
 ----------------